Remove CF notes
If people want to do this, there's nothing stopping them from doing it, but leaving it in the README implies this repo provides support for these use cases. If it breaks, it breaks, and more often than not since CF configurations are out of our control some devices will not work and some optimizations in this configuration are lost.